include curses in sheet/handlers.c move app/curses.h into app/sheet/
